Refactor commands/utils.js tests to utils/common.js.
