+++ packages/ckeditor5-table/tests/utils/structure.js
